Clarifying the first requirement
When I implemented this kata the first time, I didn't see the example (it was clipped and the text did not wrap)… curiously, my browser window width was just such that it was not obvious that the text was clipped.

In any case, I would up creating a solution that used preg_split() to explode the string on any non-integer values, and wound up solving for several of the following "problems".

I, erm, also added an Oxford comma. #becauseocd